Fix FFI numeric type.

+++ server/prototypes/rust/array/src/integration.rb
@@ -24,8 +24,8 @@ class Rust < FFI::AutoPointer
     attach_function :new,  :rust_array_new, [], Rust
     attach_function :free, :rust_array_free, [Rust], :void
                     
-    attach_function :append, :rust_array_append, [Rust, :uint32], :uint32
-    attach_function :first, :rust_array_first, [Rust], :uint32
-    attach_function :last, :rust_array_last, [Rust], :uint32
+    attach_function :append, :rust_array_append, [Rust, :uint16], :uint16
+    attach_function :first, :rust_array_first, [Rust], :uint16
+    attach_function :last, :rust_array_last, [Rust], :uint16
   end
 end
